refactor: structure persons modules
