+++ src/app/person/handlers/read.js
@@ -1,13 +1,13 @@
-import { sendJson } from '../../utils/http.js';
-import { personDatabase } from './database.js';
-import { validateUUID } from '../../utils/uuid.js';
-import { PersonNotFoundError } from './error-not-found.js';
-import { PathError } from '../../errors/path-error.js';
+import { sendJson } from '../../../utils/http.js';
+import { personDatabase } from '../service.js';
+import { validateUUID } from '../../../utils/uuid.js';
+import { PersonNotFoundError } from '../errors/error-not-found.js';
+import { PathError } from '../../../errors/path-error.js';
 
 /**
  * @type { Controller }
  */
-export const handlerRead = async (path, request, response) => {
+export const handleRead = async (path, request, response) => {
   if (path.length === 0) {
     const persons = await personDatabase.read();
     return sendJson(response, persons);
